+++ src/faas-vscode-plugin/src/ui/commands/register-commands.ts
@@ -334,6 +334,7 @@ export async function registerItemContextCommands(context: vscode.ExtensionConte
             vscode.env.openExternal(vscode.Uri.parse("https://github.com/oracle-samples/oci-vscode-toolkit/issues"));
         }),
     );
+    vscode.commands.executeCommand('setContext', 'enableFunctionsViewTitleMenus', true);
 }
 
 export async function registerGenericCommands(context: vscode.ExtensionContext) {