+++ CHANGELOG.md
@@ -7,3 +7,21 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [1.0.0]
 
 - Initial release
+
+## 1.0.1 - 2023-06-22
+
+### Added
+
+
+- Added new notification banner when Resource Manager plugin is loaded.
+
+### Changed
+
+- Changed display name and descriptions for OCI Core extension, Data Science, Functions and Resource Manager plugins.
+
+- Changed progress notification message when uploading stack changes to the RMS service in Resource Manager plugin.
+
+- Improved tooltip on stack selection in tree view in Resource Manager plugin.
+
+### Fixed
+- Enable view title menus in plugins only when tree view is loaded.